pyproject.toml
[project]
name = "mmpm"
version = "4.0.2"
description = "MMPM, the MagicMirror Package Manager CLI simplifies the installation, removal, and general maintenance of MagicMirror packages"
authors = [
  { name = "Brandon Marlowe", email = "bpmarlowe-software@protonmail.com" },
]
dependencies = [
  "Flask-Cors>=3.0.8",
  "argcomplete>=2.0.0",
  "argparse>=1.4",
  "beautifulsoup4>=4.11.1",
  "colorama>=0.4.3",
  "flask-socketio>=5.3.2",
  "flask>=2.2.0",
  "gevent>=22.10.0",
  "gunicorn>=20.1.0",
  "jinja2>=3.1.0",
  "pygments>=2.12.0",
  "python-socketio[client]>=5.7.0",
  "requests>=2.28.0",
  "yaspin>=2.3.0",
  "pip>=23.3.1",
  "gitpython>=3.1.40",
  "gevent-websocket>=0.10.1",
  "prompt-toolkit>=3.0.43",
]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }
keywords = [
  "MagicMirror magicmirror package-manager mmpm MMPM magicmirror-package-manager package manager magicmirror_package_manager",
]

[project.urls]
Homepage = "https://github.com/Bee-Mar/mmpm"
Downloads = "https://github.com/Bee-Mar/mmpm/archive/4.0.0.tar.gz"

[project.scripts]
mmpm = "mmpm.entrypoint:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
  "mypy>=1.7.0",
  "pylint>=3.0.2",
  "codecov>=2.1.13",
  "types-requests>=2.31.0.10",
  "pytest>=7.4.3",
  "faker>=20.1.0",
  "coverage>=7.3.2",
  "pdm>=2.11.1",
  "black>=23.12.0",
]

# pdm is required to make use of these (pip install pdm)
[tool.pdm.scripts]
test = "coverage run -m pytest test"
report = "coverage report"
lint = "pylint mmpm"
mypy = "mypy mmpm"
build-ui = "bash -c 'cd ui && bun run build'"
format = "bash -c 'isort . && black . --line-length=150'"

# quickstarts for dev testing
up = "pm2 start dev/ecosystem.json"
start = "pm2 start dev/ecosystem.json"

stop = "pm2 stop dev/ecosystem.json"
down = "pm2 stop dev/ecosystem.json"

rm = "pm2 delete dev/ecosystem.json"
remove = "pm2 delete dev/ecosystem.json"
delete = "pm2 delete dev/ecosystem.json"

logs = "pm2 logs mmpm"
